requirements.txt
appdirs==1.4.4
asgiref==3.3.4
boto3==1.18.64
botocore==1.21.64
distlib==0.3.1
dj-database-url==0.5.0
Django==3.1.7
django-heroku==0.3.1
django-mathfilters==1.0.0
django-phonenumber-field==5.2.0
django-storages==1.12.2
filelock==3.0.12
gunicorn==20.0.4
importlib-metadata==3.8.0
jmespath==0.10.0
phonenumbers==8.12.24
Pillow==5.2.0
psycopg2==2.8.6
python-dateutil==2.8.2
python-decouple==3.5
pytz==2021.1
s3transfer==0.5.0
six==1.15.0
sqlparse==0.4.1
typing-extensions==3.7.4.3
urllib3==1.26.7
virtualenv==20.4.3
whitenoise==5.2.0
zipp==3.4.1
